refactor: select box 분리
